// black-note-app/src/api/generated/api.schemas.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * 小黑书 API
 * 基于 Spring Boot 3 + Redis 的高性能图文博客后端
 * OpenAPI spec version: v1.0.0
 */
export interface Result {
  success?: boolean;
  errorMsg?: string;
  data?: unknown;
  total?: number;
}

export interface Voucher {
  id?: number;
  shopId?: number;
  title?: string;
  subTitle?: string;
  rules?: string;
  payValue?: number;
  actualValue?: number;
  type?: number;
  status?: number;
  stock?: number;
  beginTime?: string;
  endTime?: string;
  createTime?: string;
  updateTime?: string;
}

export interface LoginFormDto {
  phone?: string;
  code?: string;
  password?: string;
}

export interface Shop {
  id?: number;
  name?: string;
  images?: string;
  address?: string;
  area?: string;
  typeId?: number;
  avgPrice?: number;
  sold?: number;
  comments?: number;
  score?: number;
  openHours?: string;
  createTime?: string;
  updateTime?: string;
}

export interface Blog {
  id?: number;
  shopId?: number;
  userId?: number;
  icon?: string;
  name?: string;
  isLike?: boolean;
  title?: string;
  images?: string;
  content?: string;
  liked?: number;
  comments?: number;
  createTime?: string;
  updateTime?: string;
}

export interface CommentDTO {
  /** 评论内容 */
  content?: string;
  /** 父评论ID (如果是直接评论笔记，传0或不传) */
  parentId?: number;
  /** 被回复人的ID (可选，用于显示回复了谁) */
  answerId?: number;
}

export type SendCodeParams = {
/**
 * 手机号码
 */
phone: string;
};

export type UploadImageBody = {
  /** 要上传的图片文件 */
  file: Blob;
};

export type DeleteBlogImgParams = {
/**
 * 要删除的图片文件名
 */
filename: string;
};

export type QueryVoucherListParams = {
current?: number;
};

export type QueryShopListParams = {
typeId?: number;
current?: number;
};

export type QueryCommentsParams = {
current?: number;
};

export type QueryBlogByUserParams = {
/**
 * 当前页码
 */
current?: number;
/**
 * 用户ID
 */
id: number;
};

export type QueryBlogOfFollowParams = {
lastId?: number;
offset?: number;
};

export type QueryHotBlogParams = {
/**
 * 当前页码
 */
current?: number;
};